Add snapshot cleanup process
There's an extra lambda in the target account now that deletes the intermediate snapshot as soon as it's copied over to the target account.

I still need to do the final testing round.
